deployment_/src/transformer.py
from sklearn.base import BaseEstimator, TransformerMixin

class CityToRegionTransformer(BaseEstimator, TransformerMixin):
    def __init__(self, mapping):
        self.mapping = mapping

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        X = X.copy()
        X["Region"] = X["City"].map(self.mapping).fillna("Other")
        return X

class DegreeMappingTransformer(BaseEstimator, TransformerMixin):
    def __init__(self, mapping):
        self.mapping = mapping

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        X = X.copy()
        X["Degree Group"] = X["Degree"].map(self.mapping).fillna(0)
        return X

class DropColumnsTransformer(BaseEstimator, TransformerMixin):
    def __init__(self, columns):
        self.columns = columns
        
    def fit(self, X, y=None):
        return self
    
    def transform(self, X):
        return X.drop(columns=self.columns, errors="ignore")