deployment_/requirements.txt
streamlit==1.44.0
pandas
seaborn
matplotlib
plotly
pillow
numpy
datasets
scikit-learn==1.7.2